Don't make async-unsafe calls in signal handlers.
This change removes any code that is not asynchronous-safe from
signal handlers, leaving those operations to be performed in the
main flow of the program. The same strategy cannot be used for
SIGSEGV since its handler is not allowed to return normally. In
this case, the old handler was simply removed. The reset command
may be used to restore the terminal state if Rover ever crashes.