Add FAQ.md with information for common use cases.
FAQ.md is supposed  to provide extra documentation  for Rover, based
on common  queries from  users. It's  a good  place to  put detailed
information that doesn't fit the manual.

The file  rover.sh has been  removed and  an explanation for  it has
been added in FAQ.md instead. The same should be done for any helper
script that might become worth of mention in the future. That way we
don't clutter the repository with files  that aren't part of a Rover
basic setup.